added newObstacle();

+++ src/Player.java
@@ -4,7 +4,7 @@ public class Player extends Entity {
     private int score;
 
 
-    public Player(int x, int y) {
+    public Player(double x, double y) {
         super(x, y, 150, 0, 0, 500, new Image("file:../ghost.png"), 30);
         this.score = 0;
     }
